libft update

+++ libft/ft_strlcat.c
@@ -6,16 +6,17 @@
 /*   By: bde-seic <bde-seic@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/13 15:24:30 by bde-seic          #+#    #+#             */
-/*   Updated: 2022/10/16 19:27:44 by bde-seic         ###   ########.fr       */
+/*   Updated: 2022/10/18 15:21:34 by bde-seic         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
+//concatena duas strings ate ao size e retorna o tamanho total.
 size_t  ft_strlcat(char *dst, char *src, size_t size)
 {
-    int i;
-    int j;
+    size_t i;
+    size_t j;
     size_t n;
     size_t m;
