+++ libft/ft_strlcpy.c
@@ -6,15 +6,16 @@
 /*   By: bde-seic <bde-seic@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/13 12:51:59 by bde-seic          #+#    #+#             */
-/*   Updated: 2022/10/13 15:22:01 by bde-seic         ###   ########.fr       */
+/*   Updated: 2022/10/18 15:21:47 by bde-seic         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
+//copia a string src para a string dst e retorna o tamanho..
 size_t ft_strlcpy(char *dst, const char *src, size_t size)
 {
-    int i;
+    size_t i;
 
     i = 0;
     while (i < size - 1 && src[i] != 0)
